Fix: Added EventInstance policies and protected the destroy and restore methods.

+++ app/Http/Controllers/EventInstanceController.php
@@ -9,8 +9,7 @@ class EventInstanceController extends Controller
 {
     public function destroy(EventInstance $instance)
     {
-        // Authorize the request (if you have policies)
-        // $this->authorize('delete', $instance);
+        $this->authorize('destroy', $instance);
 
         $instance->delete();
 
@@ -20,6 +19,8 @@ class EventInstanceController extends Controller
     public function restore($id)
     {
         $instance = \App\Models\EventInstance::withTrashed()->findOrFail($id);
+        $this->authorize('restore', $instance);
+
         $instance->restore();
 
         return back()->with('success', 'Occurrence restored to the schedule.');
